Fix notes field.
git-svn-id: http://sandy-disaster-recovery.googlecode.com/svn/trunk@48 6201a407-fd10-26c8-02aa-e656739fd6f4

+++ sandy-disaster-recovery/key.py
@@ -60,10 +60,11 @@ def CheckAuthorization(request):
             key = Key.get_by_id(key_id)
             # Check the age of the key, and delete it
             # if it is too old.
-            age = datetime.datetime.now() - key.date
-            if age.days > 14:
-              key.delete()
-              key = None
+            if key:
+              age = datetime.datetime.now() - key.date
+              if age.days > 14:
+                key.delete()
+                key = None
             if org and key:
               if (parts[0] == key.hashOrganization(org)):
                 return org
